medical-report-analyzer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
pytesseract==0.3.10
pdf2image==0.1.21
Pillow==10.1.0
openai==1.3.0
PyYAML==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
numpy==1.24.3
scikit-learn==1.3.0
aiofiles==23.2.1
python-dotenv==1.0.0